refactor/login: Exception 요구사항에 맞게 수정
- 이메일, 비밀번호 검증 - INVALID_CREDENTIALS으로 통합
