+++ src/main/java/com/example/assignment/common/exception/custom/InvalidEmailAndPasswordException.java
@@ -2,5 +2,5 @@ package com.example.assignment.common.exception.custom;
 
 import com.example.assignment.common.exception.BusinessException;
 
-public class InvalidPasswordException extends BusinessException {
+public class InvalidEmailAndPasswordException extends BusinessException {
 }
